Removes extraneous property change notification
This property change notification causes a lot of headaches with ArrayControllers,
and it even could be a perf/UX regression. It also appears to cover an edge case that
we _shouldn't_ have, though it may pop up later. Going to talk to Igor T. about it to
get some context for why it exists