Backports rollback bugfix
If an attribute from the server is null or undefined, it won't have a key
on the internal _data property. This is a bug that was fixed in future version
of ED, it makes sense that the modified _attributes object has the keys of the
properties that we actually want to update.